+++ matdb/io.py
@@ -32,15 +32,90 @@ def order_stress(xx=None, yy=None, zz=None, yz=None, xz=None, xy=None):
 
 def symmetrize(xx=None, yy=None, zz=None, yz=None, xz=None, xy=None):
     """Returns the 3x3 stress matrix from the specified components.
-
     .. note:: the components are: xx yy zz xy yz zx.
     """
     from numpy import array
     return array([[xx, xy, xz], [xy, yy, yz], [xz, yz, zz]])
 
+ 
+def atoms_to_cfg(atm, target, config_id=None, type_map=None):
+    """Converts an :class:`matdb.atoms.Atoms` object to a cfg file.
+    
+    Args:
+        atm (matdb.atoms.Atoms): the atoms object.
+        target (str): path to the cfg file to be written.
+        config_id (str): the config id for the atoms object.
+        type_map (dict): a type map to match the species to a larger 
+          system than present in the system.
+    """
+
+    chem_syms = atm.get_chemical_symbols()
+    pos = atm.positions
+
+    if hasattr(atm, "calc") and atm.calc is not None and hasattr(atm.calc, "key"):
+        calc_name = "{0}_".format(atm.calc.key)
+    else:
+        calc_name = ""
+    
+    local_map = {}
+    for i, specs in enumerate(np.unique(chem_syms)):
+        if type_map is None:
+            local_map[specs] = i
+        else:
+            local_map[specs] =  type_map[i]
+
+    with open(target, "w+") as f:
+        f.write("BEGIN_CFG\n Size\n")
+        f.write("    {}\n".format(len(pos)))
+        f.write(" SuperCell\n")
+        # Then write out the lattice vectors.
+        lat_vecs = atm.cell
+        for i in range(3):
+            f.write("   {}\n".format("      ".join(
+                ["{0: .6f}".format(j) for j in lat_vecs[i]])))
+        
+        f.write("  ")
+
+        if "{0}force".format(calc_name) in atm.properties:
+            force = atm.properties["{0}force".format(calc_name)]
+            f.write(" AtomData:  id type       cartes_x      cartes_y      "
+                    "cartes_z           fx          fy          fz\n")
+        else:
+            f.write(" AtomData:  id type       cartes_x      cartes_y      cartes_z\n")
+        iAt = 0
+        for type, loc in zip(chem_syms, pos):
+            out_lab = local_map[type]
+            if "{0}force".format(calc_name) in atm.properties:
+                f.write("             {0}    {1}       "
+                        "{2}    {3}\n".format(iAt+1, out_lab,
+                                              "  ".join(["{0: .8f}".format(i) for i in loc]),
+                                              "    ".join(["{0: .8f}".format(i) for i in force[iAt]])))
+            else:
+                f.write("             {0}    {1}       "
+                        "{2}\n".format(iAt+1, out_lab,
+                                       "  ".join(["{0: .8f}".format(i) for i in loc])))
+            iAt += 1
+
+        if "{0}energy".format(calc_name) in atm.params:
+            f.write("  Energy\n")
+            f.write("        {0}\n".format(atm.params["{0}energy".format(calc_name)]))
+
+        if "{0}virial".format(calc_name) in atm.params:
+            virial = [atm.params["vasp_virial"][0][0], atm.params["vasp_virial"][1][1],
+                      atm.params["vasp_virial"][2][2], atm.params["vasp_virial"][2][1],
+                      atm.params["vasp_virial"][2][0], atm.params["vasp_virial"][0][1]]
+            f.write(" Stress:   xx          yy          zz          yz          xz          xy\n")
+            f.write("            {0}\n".format("    ".join(["{0: .8f}".format(i) for i in virial])))
+                        
+        if config_id is None:
+            conf_id = "{0}_{1}".format("".join(np.unique(chem_syms)),len(pos))
+        else:
+            conf_id = config_id
+        f.write(" Feature   conf_id  {}\n".format(conf_id))
+        f.write("END_CFG\n\n")
+
 def _cfgd_to_atoms(cfgd, species=None):
     """Converts a CFG dictionary to an atoms object.
-
     Args:
         cfgd (dict): of a single config extracted by :func:`cfg_to_xyz`.
         species (list): of element names corresponding to the integer species in
@@ -102,10 +177,8 @@ def _cfgd_to_atoms(cfgd, species=None):
 
 def cfg_to_xyz(cfgfile, outfile="output.xyz", config_type=None, species=None):
     """Converts MTP's CFG forrmat to XYZ.
-
     .. note:: Multiple frames in the CFG file will be converted to multiple
       frames in the XYZ file.
-
     Args:
         cfgfile (str): path to the file to convert.
         config_type (str): name of the config_type to assign to each
@@ -173,7 +246,6 @@ def vasp_to_xyz(folder, outfile="output.xyz", recalc=0,
                 config_type=None):
     """Creates an extended XYZ file for the calculated structure in
     OUTCAR for the given folder.
-
     Args:
         folder (str): path to the folder to convert.
         outfile (str): name of the XYZ file to create. The file will
@@ -239,12 +311,10 @@ def is_link(obj):
 def _unpack_obj(context, obj, lcontext=None):
     """Unpacks each item of the specified object recursively so that all
     dictionary values are visited and all list items are also visited.
-
     .. warning:: `obj` will be mutated if any value it considers turns out to be
       a link (according to :func:`is_link`). In that case, the file descriptor
       will be placed by the actual contents of the YAML file that the link
       points to.
-
     Args:
         context (str): path to the root folder where the yaml file is
           located. Needed for relative paths of file links.
@@ -282,7 +352,6 @@ def read(context, yfile):
     """Reads in the specified YAML file, following any additional file
     directives to compile a full representation of the template hierarchy for
     the root file.
-
     Args:
         context (str): path to the root folder where the yaml file is
           located. Needed for relative paths of file links.
@@ -324,7 +393,6 @@ def read(context, yfile):
 
 def save_dict_to_h5(h5file, dic, path='/'):
     """Saves a nested dictionary to an open hdf5 file.
-
     Args:
         h5file (file object): the h5 file to be saved to.
         dic (dict): the dictionary to save.
@@ -334,6 +402,13 @@ def save_dict_to_h5(h5file, dic, path='/'):
     for key, item in dic.items():
         if isinstance(item, (np.ndarray, np.int64, np.float64, str, bytes, tuple, float, int)):
             h5file[path + key] = item
+        elif isinstance(item, np.ndarray):
+            if item.ndim==1 and isinstance(item[0], np.ndarray): #pragma: no cover
+                # this chunk of code is only ever used by the unit testing suite. 
+                dt = h5py.special_dtype(vlen=np.float64)
+                h5file.create_dataset(path+key, data=item, dtype=dt)
+            else:
+                h5file[path + key] = item
         elif isinstance(item, dict):
             save_dict_to_h5(h5file, item, path + key + '/')
         elif isinstance(item, list):
@@ -350,12 +425,10 @@ def save_dict_to_h5(h5file, dic, path='/'):
 
 def load_dict_from_h5(h5file, path='/'):
     """Reads an open hdf5 file into a dictionary.
-
     Args:
         h5file (file object): the h5 file to be read.
         path (str, optional): the path within the h5 file presently being
             read. Default is '/'.
-
     Returns:
         ans (dict): a dictionary containing the contents of the h5 file.
     """